Cleanup build source code.
