+++ .github/workflows/macos-12-x86_64-x64-osx.yml
@@ -37,4 +37,4 @@ jobs:
 
       - name: Automated tests Alusus
         run: |
-          cd "$(./Tools/build.sh --target-triplet x64-osx --print-build-location)" && ctest --output-on-failure
+          bash -c "cd \"$(./Tools/build.sh --target-triplet x64-osx --print-build-location)\" && ctest --output-on-failure"
