+++ Tools/BuildSrc/alusus_common.py
@@ -1,4 +1,7 @@
+from enum import Enum
 import os
+import shutil
+import sys
 
 ALUSUS_REPO_PATH = os.path.dirname(os.path.dirname(
     os.path.dirname(os.path.realpath(__file__))))
@@ -15,3 +18,22 @@ FPM_PACKAGE_NAME = "alusus"
 FPM_PACKAGE_DESCRIPTION = "Alusus Programming Language's core compilation system and standard libraries."
 FPM_PACKAGE_MAINTAINER = "Sarmad Khalid Abdullah <sarmad@alusus.org>"
 FPM_PACKAGE_URL = "https://alusus.org"
+ALUSUS_VCPKG_REPO_PATH_ENV_VAR = "ALUSUS_VCPKG_REPO_PATH"
+
+
+class ExtendedEnum(Enum):
+    @classmethod
+    def list(cls):
+        for _ in cls:
+            return list(map(lambda c: c, cls))
+
+
+def remove_path(path):
+    if os.path.isdir(path):
+        shutil.rmtree(path, ignore_errors=True)
+    elif os.path.isfile(path):
+        os.remove(path)
+
+
+def eprint(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
