+++ Tools/BuildSrc/alusus_install_deps.py
@@ -1,19 +1,17 @@
 from __future__ import print_function
 import subprocess
-from alusus_common import ALUSUS_TOOLS_DIR
 import sys
 import os
 
-
-def eprint(*args, **kwargs):
-    print(*args, file=sys.stderr, **kwargs)
+# Alusus imports.
+import alusus_common
 
 
 if __name__ == "__main__":
     output = subprocess.run([sys.executable, "-m", "pip", "install",
-                            "-r", os.path.join(ALUSUS_TOOLS_DIR, "BuildSrc", "requirements.txt")],
+                            "-r", os.path.join(alusus_common.ALUSUS_TOOLS_DIR, "BuildSrc", "requirements.txt")],
                             stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     if output.returncode:
-        eprint(output.stderr.decode(), end='')
+        alusus_common.eprint(output.stderr.decode(), end='')
         print(output.stdout.decode(), end='')
     exit(output.returncode)
